requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
